# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "crime-intelligence-platform",
    "description": "Crime Intelligence Platform API for Karnataka State Police. Frontend contract: use /docs or /openapi.json. AI routes use Catalyst QuickML only.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/cases": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "List Cases",
        "description": "List FIRs with optional station/status/crime-head/date filters.",
        "operationId": "list_cases_api_v1_cases_get",
        "parameters": [
          {
            "name": "police_station_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Police Station Id"
            }
          },
          {
            "name": "case_status_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Case Status Id"
            }
          },
          {
            "name": "crime_major_head_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Crime Major Head Id"
            }
          },
          {
            "name": "registered_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Registered From"
            }
          },
          {
            "name": "registered_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Registered To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseMasterListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Create Case",
        "description": "Create a FIR (CaseMaster) with optional nested parties/sections.",
        "operationId": "create_case_api_v1_cases_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseMasterCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseMasterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_master_id}": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Get Case",
        "description": "Get one FIR with victims, accused, and act-sections.",
        "operationId": "get_case_api_v1_cases__case_master_id__get",
        "parameters": [
          {
            "name": "case_master_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Master Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseMasterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_master_id}/victims": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Add Victim",
        "operationId": "add_victim_api_v1_cases__case_master_id__victims_post",
        "parameters": [
          {
            "name": "case_master_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Master Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VictimCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseMasterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_master_id}/accused": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Add Accused",
        "operationId": "add_accused_api_v1_cases__case_master_id__accused_post",
        "parameters": [
          {
            "name": "case_master_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Master Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccusedCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseMasterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_master_id}/act-sections": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Add Act Section",
        "operationId": "add_act_section_api_v1_cases__case_master_id__act_sections_post",
        "parameters": [
          {
            "name": "case_master_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Master Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActSectionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseMasterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccusedCreate": {
        "properties": {
          "accused_name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Accused Name"
          },
          "age_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Year"
          },
          "gender_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender Id"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          }
        },
        "type": "object",
        "required": [
          "accused_name"
        ],
        "title": "AccusedCreate"
      },
      "AccusedRead": {
        "properties": {
          "accused_master_id": {
            "type": "integer",
            "title": "Accused Master Id"
          },
          "case_master_id": {
            "type": "integer",
            "title": "Case Master Id"
          },
          "accused_name": {
            "type": "string",
            "title": "Accused Name"
          },
          "age_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Year"
          },
          "gender_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender Id"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          }
        },
        "type": "object",
        "required": [
          "accused_master_id",
          "case_master_id",
          "accused_name"
        ],
        "title": "AccusedRead"
      },
      "ActSectionCreate": {
        "properties": {
          "act_id": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Act Id"
          },
          "section_id": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Section Id"
          },
          "act_order_id": {
            "type": "integer",
            "title": "Act Order Id",
            "default": 1
          },
          "section_order_id": {
            "type": "integer",
            "title": "Section Order Id",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "act_id",
          "section_id"
        ],
        "title": "ActSectionCreate"
      },
      "ActSectionRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "case_master_id": {
            "type": "integer",
            "title": "Case Master Id"
          },
          "act_id": {
            "type": "string",
            "title": "Act Id"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "act_order_id": {
            "type": "integer",
            "title": "Act Order Id"
          },
          "section_order_id": {
            "type": "integer",
            "title": "Section Order Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_master_id",
          "act_id",
          "section_id",
          "act_order_id",
          "section_order_id"
        ],
        "title": "ActSectionRead"
      },
...
      "CaseMasterCreate": {
        "properties": {
          "crime_no": {
            "type": "string",
            "maxLength": 18,
            "minLength": 18,
            "title": "Crime No"
          },
          "case_no": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Case No"
          },
          "crime_registered_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crime Registered Date"
          },
          "police_station_id": {
            "type": "integer",
            "title": "Police Station Id"
          },
          "case_category_id": {
            "type": "integer",
            "title": "Case Category Id"
          },
          "gravity_offence_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gravity Offence Id"
          },
          "crime_major_head_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crime Major Head Id"
          },
          "crime_minor_head_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crime Minor Head Id"
          },
          "case_status_id": {
            "type": "integer",
            "title": "Case Status Id"
          },
          "court_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Court Id"
          },
          "incident_from_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident From Date"
          },
          "incident_to_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident To Date"
          },
          "info_received_ps_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Info Received Ps Date"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "brief_facts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brief Facts"
          },
          "victims": {
            "items": {
              "$ref": "#/components/schemas/VictimCreate"
            },
            "type": "array",
            "title": "Victims"
          },
          "accused": {
            "items": {
              "$ref": "#/components/schemas/AccusedCreate"
            },
            "type": "array",
            "title": "Accused"
          },
          "act_sections": {
            "items": {
              "$ref": "#/components/schemas/ActSectionCreate"
            },
            "type": "array",
            "title": "Act Sections"
          }
        },
        "type": "object",
        "required": [
          "crime_no",
          "case_no",
          "police_station_id",
          "case_category_id",
          "case_status_id"
        ],
        "title": "CaseMasterCreate"
      },
      "CaseMasterDetail": {
        "properties": {
          "crime_no": {
            "type": "string",
            "maxLength": 18,
            "minLength": 18,
            "title": "Crime No"
          },
          "case_no": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Case No"
          },
          "crime_registered_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crime Registered Date"
          },
          "police_station_id": {
            "type": "integer",
            "title": "Police Station Id"
          },
          "case_category_id": {
            "type": "integer",
            "title": "Case Category Id"
          },
          "gravity_offence_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gravity Offence Id"
          },
          "crime_major_head_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crime Major Head Id"
          },
          "crime_minor_head_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crime Minor Head Id"
          },
          "case_status_id": {
            "type": "integer",
            "title": "Case Status Id"
          },
          "court_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Court Id"
          },
          "incident_from_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident From Date"
          },
          "incident_to_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident To Date"
          },
          "info_received_ps_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Info Received Ps Date"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "brief_facts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brief Facts"
          },
          "case_master_id": {
            "type": "integer",
            "title": "Case Master Id"
          },
          "victims": {
            "items": {
              "$ref": "#/components/schemas/VictimRead"
            },
            "type": "array",
            "title": "Victims"
          },
          "accused": {
            "items": {
              "$ref": "#/components/schemas/AccusedRead"
            },
            "type": "array",
            "title": "Accused"
          },
          "act_sections": {
            "items": {
              "$ref": "#/components/schemas/ActSectionRead"
            },
            "type": "array",
            "title": "Act Sections"
          }
        },
        "type": "object",
        "required": [
          "crime_no",
          "case_no",
          "police_station_id",
          "case_category_id",
          "case_status_id",
          "case_master_id"
        ],
        "title": "CaseMasterDetail"
      },
      "CaseMasterListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseMasterRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "CaseMasterListResponse"
      },
      "CaseMasterRead": {
        "properties": {
          "crime_no": {
            "type": "string",
            "maxLength": 18,
            "minLength": 18,
            "title": "Crime No"
          },
          "case_no": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Case No"
          },
          "crime_registered_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crime Registered Date"
          },
          "police_station_id": {
            "type": "integer",
            "title": "Police Station Id"
          },
          "case_category_id": {
            "type": "integer",
            "title": "Case Category Id"
          },
          "gravity_offence_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gravity Offence Id"
          },
          "crime_major_head_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crime Major Head Id"
          },
          "crime_minor_head_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crime Minor Head Id"
          },
          "case_status_id": {
            "type": "integer",
            "title": "Case Status Id"
          },
          "court_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Court Id"
          },
          "incident_from_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident From Date"
          },
          "incident_to_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident To Date"
          },
          "info_received_ps_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Info Received Ps Date"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "brief_facts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brief Facts"
          },
          "case_master_id": {
            "type": "integer",
            "title": "Case Master Id"
          }
        },
        "type": "object",
        "required": [
          "crime_no",
          "case_no",
          "police_station_id",
          "case_category_id",
          "case_status_id",
          "case_master_id"
        ],
        "title": "CaseMasterRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VictimCreate": {
        "properties": {
          "victim_name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Victim Name"
          },
          "age_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Year"
          },
          "gender_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender Id"
          },
          "victim_police": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Victim Police",
            "default": "0"
          }
        },
        "type": "object",
        "required": [
          "victim_name"
        ],
        "title": "VictimCreate"
      },
      "VictimRead": {
        "properties": {
          "victim_master_id": {
            "type": "integer",
            "title": "Victim Master Id"
          },
          "case_master_id": {
            "type": "integer",
            "title": "Case Master Id"
          },
          "victim_name": {
            "type": "string",
            "title": "Victim Name"
          },
          "age_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Year"
          },
          "gender_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender Id"
          },
          "victim_police": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Victim Police"
          }
        },
        "type": "object",
        "required": [
          "victim_master_id",
          "case_master_id",
          "victim_name"
        ],
        "title": "VictimRead"
      }
    }
  },
...
}